fix: minichef hooks folder
