+++ src/components/Pools/DetailModal/DetailView/index.tsx
@@ -6,7 +6,7 @@ import { Box } from 'src/components';
 import { usePair } from 'src/data/Reserves';
 import { useChainId } from 'src/hooks';
 import { PangoChefInfo } from 'src/state/ppangoChef/types';
-import { useGetPoolDollerWorth } from 'src/state/pstake/hooks';
+import { useGetPoolDollerWorth } from 'src/state/pstake/hooks/common';
 import { DoubleSideStakingInfo } from 'src/state/pstake/types';
 import { unwrappedToken } from 'src/utils/wrappedCurrency';
 import EarnWidget from '../../EarnWidget';
